requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
humanfriendly==4.17
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector-python==8.0.13
numpy==1.15.3
pandas==0.23.4
protobuf==3.6.1
pylint==2.1.1
python-dateutil==2.7.4
pytz==2018.6
requests==2.20.0
selenium==3.14.1
six==1.11.0
urllib3==1.24
wrapt==1.10.11
